fix decoding of zigzag32 negative values once cast to a 64-bit 'int' type
the bug dates back to a [mis]feature of the upstream google code: since
their decoder never puts the result of a DecodeZigzag32() in anything
but an 'int32' type it didnt matter what the upper 32 bits of the result
where. They had them hardcoded to 0. Aka the signed 32-bit result was
embedded in the lower 32 bits of the unsigned 64-bit result.

That breaks in my code, where I might cast the result to an 'int' on
targets where 'int' is 64-bits. The fix is to return a signed 64-bit
result. Then everyone is happy. That does require an adjustment to
all_test.go, since the raw 64-but result changed.

+++ protobuf3/all_test.go
@@ -191,7 +191,11 @@ func TestNumericPrimitives(t *testing.T) {
 		if e != nil {
 			t.Fatal("DecodeZigzag32")
 		}
-		if x != uint64(uint32(i32)) {
+		// NOTE WELL google's DecodeZigzag32() returns a signed 32-bit inside an unsigned 64-bit (bits 32-63 are 0).
+		// My DecodeZigzag32() returns a signed 64-bit. So this comparison is slightly different than in the proto/all_test.go
+		// (mine is different because unlike google I might be casting the result to an 'int', as well as 'int32', and 'int'
+		// might turn out to be 64 bits)
+		if x != uint64(i32) {
 			t.Fatal("zigzag32 decode fail:", i32, x)
 		}
